[TASK-002] Renaming the listing page file
